// web/src/api/models/feedback.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ApeRAG API
 * ApeRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Feedback
 */
export interface Feedback {
    /**
     * 
     * @type {string}
     * @memberof Feedback
     */
    'type'?: FeedbackTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Feedback
     */
    'tag'?: FeedbackTagEnum;
    /**
     * 
     * @type {string}
     * @memberof Feedback
     */
    'message'?: string;
}

export const FeedbackTypeEnum = {
    good: 'good',
    bad: 'bad'
} as const;

export type FeedbackTypeEnum = typeof FeedbackTypeEnum[keyof typeof FeedbackTypeEnum];
export const FeedbackTagEnum = {
    Harmful: 'Harmful',
    Unsafe: 'Unsafe',
    Fake: 'Fake',
    Unhelpful: 'Unhelpful',
    Other: 'Other'
} as const;

export type FeedbackTagEnum = typeof FeedbackTagEnum[keyof typeof FeedbackTagEnum];